# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PalComTech CRM API",
    "version": "1.0.0",
    "description": "API WhatsApp CRM. Autentikasi pakai API key (Pengaturan → API Key). Kirim header `X-API-Key` atau `Authorization: Bearer <key>` di setiap request."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "ApiKeyAuth": [] }, { "BearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "X-API-Key" },
      "BearerAuth": { "type": "http", "scheme": "bearer" }
    }
  },
  "paths": {
...
    "/keys": {
      "get": { "summary": "Daftar API key (khusus admin)", "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Buat API key (khusus admin)",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "label": { "type": "string", "example": "App Android" } } } } } },
        "responses": { "200": { "description": "{ key }" } }
      }
    },
    "/keys/{key}": {
      "delete": {
        "summary": "Hapus API key (khusus admin)",
        "parameters": [{ "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}